Renamed redux files
