+++ src/js/containers/submissionsTable/FilterBarContainer.jsx
@@ -9,7 +9,7 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { isEqual } from 'lodash';
 
-import * as dashboardFilterActions from 'redux/actions/submissionsTable/dashboardFilterActions';
+import * as dashboardFilterActions from 'redux/actions/submissionsTable/stagedFiltersActions';
 
 import FilterBar from 'components/submissionsTable/filters/FilterBar';
