+++ src/js/redux/reducers/submissionsTable/stagedFiltersReducer.js
